requirements.txt
web3==6.0.0
paho-mqtt==1.6.1
py-ecc==6.0.0
